day 4
commit 7: I have started to try and implement the cloud storage. It is
not working so far but I am watching the video and trying to get it
working. Also I can’t pass objects from the pokemon table to the window
that it opens with “this.” for some reason. I think it has to do with
the fact that it is not in an array. I have tried multiple times to get
it working with no avail.