day 4
all of my code was reset to a none working version that I committed 4
hours ago… I tried as best I could to get it back to this point before
the deadline, but I am having an error that I have no idea why it is
happening. I have an infinite loop of one of my api pulls that is
printing to the screen and I have no idea why. I am done